Removed socials styling from Navigation, now passing open boolean to Socials which adds class of navOpen for styling.
