+++ src/components/Socials/data.js
@@ -1,0 +1,14 @@
+const data = [
+	{
+		label: "Facebook",
+		destination: "https://www.facebook.com/"
+	}, {
+		label: "Twitter",
+		destination: "https://www.twitter.com/"
+	}, {
+		label: "Instagram",
+		destination: "https://www.instagram.com/"
+	}
+];
+
+export default data;
