+++ src/components/Socials/index.js
@@ -1,5 +1,6 @@
 import s from "./styles.scss";
+import data from "./data.js";
 import RENDER from "./renderUtils.jsx";
 
-export { RENDER, s };
+export { RENDER, s, data };
 export { default } from "./Socials.jsx";
